+++ ormar/fields/foreign_key.py
@@ -162,7 +162,7 @@ class ForeignKeyField(BaseField):
 
     @classmethod
     def _extract_model_from_sequence(
-        cls, value: List, child: "Model", to_register: bool
+        cls, value: List, child: "Model", to_register: bool, relation_name: str
     ) -> List["Model"]:
         """
         Takes a list of Models and registers them on parent.
@@ -180,13 +180,18 @@ class ForeignKeyField(BaseField):
         :rtype: List["Model"]
         """
         return [
-            cls.expand_relationship(val, child, to_register)  # type: ignore
+            cls.expand_relationship(  # type: ignore
+                value=val,
+                child=child,
+                to_register=to_register,
+                relation_name=relation_name,
+            )
             for val in value
         ]
 
     @classmethod
     def _register_existing_model(
-        cls, value: "Model", child: "Model", to_register: bool
+        cls, value: "Model", child: "Model", to_register: bool, relation_name: str
     ) -> "Model":
         """
         Takes already created instance and registers it for parent.
@@ -204,12 +209,12 @@ class ForeignKeyField(BaseField):
         :rtype: Model
         """
         if to_register:
-            cls.register_relation(value, child)
+            cls.register_relation(model=value, child=child, relation_name=relation_name)
         return value
 
     @classmethod
     def _construct_model_from_dict(
-        cls, value: dict, child: "Model", to_register: bool
+        cls, value: dict, child: "Model", to_register: bool, relation_name: str
     ) -> "Model":
         """
         Takes a dictionary, creates a instance and registers it for parent.
@@ -231,12 +236,12 @@ class ForeignKeyField(BaseField):
             value["__pk_only__"] = True
         model = cls.to(**value)
         if to_register:
-            cls.register_relation(model, child)
+            cls.register_relation(model=model, child=child, relation_name=relation_name)
         return model
 
     @classmethod
     def _construct_model_from_pk(
-        cls, value: Any, child: "Model", to_register: bool
+        cls, value: Any, child: "Model", to_register: bool, relation_name: str
     ) -> "Model":
         """
         Takes a pk value, creates a dummy instance and registers it for parent.
@@ -263,11 +268,13 @@ class ForeignKeyField(BaseField):
             )
         model = create_dummy_instance(fk=cls.to, pk=value)
         if to_register:
-            cls.register_relation(model, child)
+            cls.register_relation(model=model, child=child, relation_name=relation_name)
         return model
 
     @classmethod
-    def register_relation(cls, model: "Model", child: "Model") -> None:
+    def register_relation(
+        cls, model: "Model", child: "Model", relation_name: str
+    ) -> None:
         """
         Registers relation between parent and child in relation manager.
         Relation manager is kep on each model (different instance).
@@ -281,12 +288,20 @@ class ForeignKeyField(BaseField):
         :type child: Model class
         """
         model._orm.add(
-            parent=model, child=child, child_name=cls.related_name, virtual=cls.virtual
+            parent=model,
+            child=child,
+            child_name=cls.related_name or child.get_name() + "s",
+            virtual=cls.virtual,
+            relation_name=relation_name,
         )
 
     @classmethod
     def expand_relationship(
-        cls, value: Any, child: Union["Model", "NewBaseModel"], to_register: bool = True
+        cls,
+        value: Any,
+        child: Union["Model", "NewBaseModel"],
+        to_register: bool = True,
+        relation_name: str = None,
     ) -> Optional[Union["Model", List["Model"]]]:
         """
         For relations the child model is first constructed (if needed),
@@ -316,5 +331,5 @@ class ForeignKeyField(BaseField):
 
         model = constructors.get(  # type: ignore
             value.__class__.__name__, cls._construct_model_from_pk
-        )(value, child, to_register)
+        )(value, child, to_register, relation_name)
         return model